Roles and permissions

+++ app/Http/Controllers/TasksController.php
@@ -6,11 +6,15 @@ use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Models\Task;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class TasksController extends Controller
 {
     public function index()
     {
+        abort_if(Gate::denies('task_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $tasks = Task::all();
 
         return view('tasks.index', compact('tasks'));
@@ -18,6 +22,8 @@ class TasksController extends Controller
 
     public function create()
     {
+        abort_if(Gate::denies('task_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('tasks.create');
     }
 
@@ -30,11 +36,15 @@ class TasksController extends Controller
 
     public function show(Task $task)
     {
+        abort_if(Gate::denies('task_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('tasks.show', compact('task'));
     }
 
     public function edit(Task $task)
     {
+        abort_if(Gate::denies('task_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('tasks.edit', compact('task'));
     }
 
@@ -47,6 +57,8 @@ class TasksController extends Controller
 
     public function destroy(Task $task)
     {
+        abort_if(Gate::denies('task_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $task->delete();
 
         return redirect()->route('tasks.index');
